refactored some package names and interfaces
