+++ src/main/java/DAO/FoodDAO.java
@@ -1,6 +1,6 @@
 package DAO;
 
-import Beans.Food;
+import DTO.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
