+++ src/main/java/DAO/UnhealthyToHealthyDAO.java
@@ -1,11 +1,10 @@
 package DAO;
 
-import Beans.UnhealthyToHealthy;
+import DTO.UnhealthyToHealthy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UnhealthyToHealthyDAO extends JpaRepository<Beans.UnhealthyToHealthy,Long>{
-    List<UnhealthyToHealthy> findByUnhealthyFoodId(Long unhealthyFoodId);
+public interface UnhealthyToHealthyDAO extends JpaRepository<DTO.UnhealthyToHealthy,Long>{
 }
